requirements.txt
pytesseract
pillow
openai
